redditrepostsleuth/requirements.txt
praw
sqlalchemy
alembic
pymysql
pillow
requests
distance
celery[redis]
numpy
dataclasses
flower
eventlet
ImageHash
annoy
RedLock
influxdb
requests
gensim
ffmpeg-python
youtube-dl
wavio
pydub==0.20.0
numpy
matplotlib
scipy
numpy
